+++ packages/rb-apps/src/index.ts
@@ -6,6 +6,8 @@ export * from './types';
 export * from './AppRegistry';
 export * from './stores/filesStore';
 export * from './stores/fileAssociationsStore';
+export * from './stores/fileSystemStore';
+export * from './apps/files/fileActionTargets';
 export * from './examples';
 
 export { TerminalApp } from './apps/TerminalApp';